@@ -3,44 +3,15 @@ module.exports = {
   env: {
     node: true
   },
-  extends: [
-    'plugin:vue/recommended',
-    '@vue/standard'
+  'extends': [
+    'plugin:vue/vue3-essential',
+    'eslint:recommended'
   ],
   parserOptions: {
     parser: '@babel/eslint-parser'
   },
   rules: {
-    'arrow-parens': ['error', 'as-needed', { requireForBlockBody: true }],
-    'import/no-extraneous-dependencies': 'off',
-    'indent': ['error', 'tab'],
     'no-console': process.env.NODE_ENV === 'production' ? 'warn' : 'off',
-    'no-debugger': process.env.NODE_ENV === 'production' ? 'warn' : 'off',
-    'no-tabs': 'off',
-    'prefer-destructuring': ['error', {
-      'object': true,
-      'array': false,
-    }],
-    'vue/html-indent': ['error', 'tab'],
-    'vue/max-attributes-per-line': ['error', {
-      'singleline': 3,
-      'multiline': {
-        'max': 1,
-        'allowFirstLine': false
-      }
-    }],
-    'vue/script-indent': ['error', 'tab', {
-      'baseIndent': 1,
-      'switchCase': 1,
-    }],
-    'vue/singleline-html-element-content-newline': 'off',
-  },
-  overrides: [
-    {
-      files: ['*.vue'],
-      rules: {
-        indent: 'off',
-      },
-    },
-  ],
-};
+    'no-debugger': process.env.NODE_ENV === 'production' ? 'warn' : 'off'
+  }
+}
